+++ packages/material-ui/src/Slider/index.js
@@ -1,4 +1,5 @@
-export { default } from './SliderStyled';
+export { default } from './Slider';
+export { default as SliderBase } from './SliderBase';
 export {
   SliderRoot,
   SliderMark,
@@ -7,4 +8,4 @@ export {
   SliderTrack,
   SliderThumb,
   SliderValueLabel,
-} from './SliderStyled';
+} from './Slider';
